Rename remaining actions
