+++ cmd/flowrunner/testdata/flows/all_actions.json
@@ -85,7 +85,7 @@
                         },
                         {
                             "uuid": "ac110f56-a66c-4462-921c-b2c6d1c6dadb",
-                            "type": "send_msg",
+                            "type": "send_broadcast",
                             "urns": [
                                 "tel:+12065551212"
                             ],
@@ -93,7 +93,7 @@
                         },
                         {
                             "uuid": "c0057fd9-be0a-43ea-91df-5c18e14f2c59",
-                            "type": "send_msg",
+                            "type": "send_broadcast",
                             "groups": [
                                 {
                                     "uuid": "2aad21f6-30b7-42c5-bd7f-1b720c154817",
@@ -117,18 +117,18 @@
                         },
                         {
                             "uuid": "f01d693b-2af2-49fb-9e38-146eb00937e9",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Hi @contact.name, are you ready to complete today's survey?"
                         },
                         {
                             "uuid": "d98c1e02-69df-4f95-8b89-8587a57ae0c3",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "This is a message to each of @contact.name's urns.",
                             "all_urns": true
                         },
                         {
                             "uuid": "62a30ab4-d73c-447d-a989-39c49115153e",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "This is a reply with attachments and quick replies",
                             "attachments": [
                                 "image/jpeg:http://s3.amazon.com/bucket/test_en.jpg?a=@contact.fields.state"
@@ -140,7 +140,7 @@
                         },
                         {
                             "uuid": "5508e6a7-26ce-4b3b-b32e-bb4e2e614f5d",
-                            "type": "save_flow_result",
+                            "type": "set_run_result",
                             "name": "Gender",
                             "value": "m",
                             "category": "Male"
