+++ cmd/flowrunner/testdata/flows/brochure.json
@@ -11,7 +11,7 @@
                     "uuid": "32bc60ad-5c86-465e-a6b8-049c44ecce49",
                     "actions": [
                         {
-                            "type": "reply",
+                            "type": "send_msg",
                             "uuid": "9d9290a7-3713-4c22-8821-4af0a64c0821",
                             "text": "Hi! What is your name?"
                         }
@@ -82,7 +82,7 @@
                         },
                         {
                             "uuid": "605e3486-503d-481c-94f7-cd553f196a8a",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Great, you are @contact.name, thanks for joining!"
                         }
                     ]
