+++ cmd/flowrunner/testdata/flows/date_parse.json
@@ -12,7 +12,7 @@
                     "actions": [
                         {
                             "uuid": "e97cd6d5-3354-4dbd-85bc-6c1f87849eec",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Hi @contact.name! When were you born, enter in format yyyy.MM.dd"
                         }
                     ],
@@ -60,7 +60,7 @@
                         },
                         {
                             "uuid": "d2a4052a-3fa9-4608-ab3e-5b9631440447",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Awesome, you were born on @(FORMAT_DATE(run.results.birth_date, \"MM-dd-yyyy\")) at @(FORMAT_DATE(run.results.birth_date, \"HH:mm\"))"
                         }
                     ]
