+++ cmd/flowrunner/testdata/flows/subflow.json
@@ -12,7 +12,7 @@
                     "actions": [
                         {
                             "uuid": "49f6c984-620f-4d9b-98c4-8ead1d1ef4f6",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "This is the parent flow"
                         },
                         {
@@ -61,7 +61,7 @@
                     "actions": [
                         {
                             "uuid": "5d51eae6-be0f-4cc7-9402-150aa1ed80a1",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Flow succeeded, they said @child.results.name"
                         }
                     ]
@@ -71,7 +71,7 @@
                     "actions": [
                         {
                             "uuid": "d80b2a5c-3b5c-47cd-b6ea-2f59bf2bb477",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Flow failed"
                         }
                     ]
@@ -92,7 +92,7 @@
                     "actions": [
                         {
                             "uuid": "e5a03dde-3b2f-4603-b5d0-d927f6bcc361",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "What is your name?"
                         }
                     ],
@@ -130,7 +130,7 @@
                     "actions": [
                         {
                             "uuid": "d63929fe-e999-42ef-abf1-4b281f58891e",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Got it!"
                         }
                     ]
