+++ cmd/flowrunner/testdata/flows/triggered.json
@@ -13,7 +13,7 @@
                     "actions": [
                         {
                             "uuid": "e97cd6d5-3354-4dbd-85bc-6c1f87849eec",
-                            "type": "reply",
+                            "type": "send_msg",
                             "text": "Hi @contact.name you were started in this flow by @parent.contact.name from the '@parent.flow' flow. He is from @parent.contact.fields.state and is aged @parent.results.age."
                         }
                     ]
