+++ cmd/flowserver/static/start.json
@@ -13,7 +13,7 @@
                         "actions": [
                             {
                                 "uuid": "accca030-7963-4660-92e9-5b86d5f21d05",
-                                "type": "reply",
+                                "type": "send_msg",
                                 "text": "You said '@run.input.text'"
                             }
                         ]
