+++ flows/actions/set_run_result.go
@@ -7,10 +7,10 @@ import (
 	"github.com/nyaruka/goflow/utils"
 )
 
-// TypeSaveFlowResult is our type for the save result action
-const TypeSaveFlowResult string = "save_flow_result"
+// TypeSetRunResult is the type for the set run result action
+const TypeSetRunResult string = "set_run_result"
 
-// SaveFlowResultAction can be used to save a result for a flow. The result will be available in the context
+// SetRunResultAction can be used to save a result for a flow. The result will be available in the context
 // for the run as @run.results.[name]. The optional category can be used as a way of categorizing results,
 // this can be useful for reporting or analytics.
 //
@@ -20,15 +20,15 @@ const TypeSaveFlowResult string = "save_flow_result"
 // ```
 //   {
 //     "uuid": "8eebd020-1af5-431c-b943-aa670fc74da9",
-//     "type": "save_flow_result",
+//     "type": "set_run_result",
 //     "name": "Gender",
 //     "value": "m",
 //     "category": "Male"
 //   }
 // ```
 //
-// @action save_flow_result
-type SaveFlowResultAction struct {
+// @action set_run_result
+type SetRunResultAction struct {
 	BaseAction
 	Name     string `json:"name" validate:"required"`
 	Value    string `json:"value" validate:"required"`
@@ -36,15 +36,15 @@ type SaveFlowResultAction struct {
 }
 
 // Type returns the type of this action
-func (a *SaveFlowResultAction) Type() string { return TypeSaveFlowResult }
+func (a *SetRunResultAction) Type() string { return TypeSetRunResult }
 
 // Validate validates our action is valid and has all the assets it needs
-func (a *SaveFlowResultAction) Validate(assets flows.SessionAssets) error {
+func (a *SetRunResultAction) Validate(assets flows.SessionAssets) error {
 	return nil
 }
 
 // Execute runs this action
-func (a *SaveFlowResultAction) Execute(run flows.FlowRun, step flows.Step, log flows.EventLog) error {
+func (a *SetRunResultAction) Execute(run flows.FlowRun, step flows.Step, log flows.EventLog) error {
 	// get our localized value if any
 	template := run.GetText(utils.UUID(a.UUID()), "value", a.Value)
 	value, err := excellent.EvaluateTemplateAsString(run.Environment(), run.Context(), template, false)
